+++ tests/switcheroo/ssh/integration/test_file_metric_publishing.py
@@ -1,6 +1,8 @@
 from switcheroo.ssh.data_org.publisher import FileKeyPublisher
 from switcheroo.ssh.data_org.publisher.s3 import S3KeyPublisher
 from metric_system.functions.file_metric_publisher import FileMetricPublisher
+#from tests.test_cloud_watch.retrievers import FileMetricRetriever
+#from tests.test_cloud_watch.retrievers.cloudwatch import AWSMetricRetriever
 
 def test_file_metrics_publish_with_file_key_publisher(
         file_key_publisher: FileKeyPublisher,
@@ -8,8 +10,10 @@ def test_file_metrics_publish_with_file_key_publisher(
         some_host: str,
         some_name: str,
 ):
+    #publish the SSH keys and metrics
     file_key_publisher.publish_key(some_host, some_name, file_metric_publisher)
     #check if metrics published
+
 
 def test_file_metrics_publish_with_s3_key_publisher(
         s3_key_publisher: S3KeyPublisher,
@@ -17,5 +21,6 @@ def test_file_metrics_publish_with_s3_key_publisher(
         some_host: str,
         some_name: str,
 ):
+    #publish the SSH keys and metrics
     s3_key_publisher.publish_key(some_host, some_name, file_metric_publisher)
     #check if metrics published
